backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqladmin==0.16.1
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis==5.0.1
hiredis==2.3.2

# Celery and message queue
celery==5.3.6
kombu==5.3.5
amqp==5.2.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
bcrypt==4.1.2
python-dotenv==1.0.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0
cryptography==42.0.0

# HTTP requests
httpx==0.26.0
aiohttp==3.9.1

# Translation services
#googletrans==4.0.0rc1
deep-translator==1.11.4

# Email
fastapi-mail==1.4.1
emails==0.6

# File handling
python-magic==0.4.27
Pillow==10.2.0
aiofiles==23.2.1

# Payment gateways (Uzbek)
pydantic==2.5.3

# Validation
email-validator==2.1.0

# Utilities
python-slugify==8.0.1
python-dateutil==2.8.2
pytz==2023.3.post1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
faker==22.0.0
httpx==0.26.0

# Development
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Monitoring
prometheus-client==0.19.0

# WebSockets
websockets==12.0
itsdangerous>=2.0.0
